Revert "Move *new* `errata_details` endpoint from `/v1` to `/v2`"
This reverts commit 69243e2c7e2d849cadd53b142f81eb6a7e59b55b.
