+++ lib/trento_web/views/v1/suse_manager_view.ex
@@ -63,9 +63,15 @@ defmodule TrentoWeb.V1.SUSEManagerView do
 
   def render("patches_for_packages.json", %{patches: patches}), do: %{patches: patches}
 
-  def render("errata_details.json", %{errata_details: errata_details = %{errataFrom: errataFrom}}),
-    do:
-      errata_details
-      |> Map.drop([:errataFrom])
-      |> Map.put(:errata_from, errataFrom)
+  def render("errata_details.json", %{
+        errata_details: errata_details = %{errataFrom: errataFrom},
+        fixes: fixes
+      }),
+      do: %{
+        errata_details:
+          errata_details
+          |> Map.drop([:errataFrom])
+          |> Map.put(:errata_from, errataFrom),
+        fixes: fixes
+      }
 end
