+++ test/trento_web/views/v1/suse_manager_view_test.exs
@@ -23,13 +23,21 @@ defmodule TrentoWeb.V1.SUSEManagerViewTest do
   describe "renders errata_details.json" do
     test "should render relevant fields" do
       %{errataFrom: errata_from} = errata_details = build(:errata_details)
+      fixes = build(:bugzilla_fix)
 
       errata_details_sans_errata_from = Map.delete(errata_details, :errataFrom)
 
-      assert Map.put(errata_details_sans_errata_from, :errata_from, errata_from) ==
+      expected_errata_details =
+        Map.put(errata_details_sans_errata_from, :errata_from, errata_from)
+
+      assert %{
+               errata_details: ^expected_errata_details,
+               fixes: ^fixes
+             } =
                render(SUSEManagerView, "errata_details.json", %{
                  errata_details:
-                   Map.put(errata_details_sans_errata_from, :errataFrom, errata_from)
+                   Map.put(errata_details_sans_errata_from, :errataFrom, errata_from),
+                 fixes: fixes
                })
     end
   end
